+++ packages/expo-battery/src/ExpoBattery.web.ts
@@ -52,7 +52,7 @@ export default {
   async getBatteryStateAsync(): Promise<BatteryState> {
     const batteryManager = await getBatteryManagerAsync();
     if (!batteryManager) return BatteryState.UNKNOWN;
-    return batteryManager.charging ? BatteryState.CHARGING : BatteryState.UNPLUGGED;
+    return getBatteryState(batteryManager.charging, batteryManager.level);
   },
 
   async startObserving() {
@@ -70,18 +70,37 @@ export default {
   },
 };
 
-function onChargingChange(this: BatteryManager): void {
-  const batteryState = this.charging ? BatteryState.CHARGING : BatteryState.UNPLUGGED;
+let lastReportedState: BatteryState = BatteryState.UNKNOWN;
+
+function getBatteryState(isCharging: boolean, level: number): BatteryState {
+  return isCharging
+    ? level >= 1.0
+      ? BatteryState.FULL
+      : BatteryState.CHARGING
+    : BatteryState.UNPLUGGED;
+}
+
+function emitStateChange(isCharging: boolean, level: number) {
+  const batteryState = getBatteryState(isCharging, level);
+  // prevent sending the same state change twice.
+  if (batteryState === lastReportedState) return;
+  lastReportedState = batteryState;
   emitter.emit('Expo.batteryStateDidChange', { batteryState });
+}
+
+function onChargingChange(this: BatteryManager): void {
+  emitStateChange(this.charging, this.level);
 }
 
 function onLevelChange(this: BatteryManager): void {
   const batteryLevel = this.level;
+  // update the state as well in case the state changed to full.
+  emitStateChange(this.charging, this.level);
   emitter.emit('Expo.batteryLevelDidChange', { batteryLevel });
 }
 
 async function getBatteryManagerAsync(): Promise<BatteryManager | null> {
-  if (!Platform.isDOMAvailable) return null;
+  if (Platform.isDOMAvailable === false) return null;
   if ('getBattery' in navigator) {
     // @ts-ignore
     return await navigator.getBattery();